refactor: move skip methods to abstract parser
Move the PostgreSQL skip methods from the PostgreSQL parser to the
abstract parser. This is step 1 in refactoring the GoogleSQL and
PostgreSQL parser so they can share more code. The eventual goal is to
allow the GoogleSQL parser to be able to handle SQL string without
having to remove the comments from the string first.
